+++ Northwind.EntityModels.Sqlite/Product.cs
@@ -31,7 +31,7 @@ public partial class Product
     public string? QuantityPerUnit { get; set; }
 
     [Column(TypeName = "money")]
-    public double? UnitPrice { get; set; }
+    public decimal? UnitPrice { get; set; }
 
     [Column(TypeName = "smallint")]
     public short? UnitsInStock { get; set; }
